ErrorManager: direct copy loggableValue
